fixing up the project structure
