+++ src/game/CMakeLists.txt
@@ -1,9 +1,17 @@
 set(GAME
+   src/game/GameIncludes.h
    src/game/GameTypes.h
+   src/game/GameObjectManager.h
+   src/game/GameObjectManager.c
    src/game/Game.h
    src/game/Game.c
    src/game/GameConstants.h
    src/game/GameConstants.c
+   src/game/Physics.h
+   src/game/Physics.c
+
+
+
    src/game/GameObjects/Camera.h
    src/game/GameObjects/Camera.c
    src/game/GameObjects/Cube.h
@@ -16,5 +24,6 @@ set(GAME
    src/game/GameObjects/GameGeometry/DebugLevel/Floors.c
    src/game/GameObjects/GameGeometry/DebugLevel/Walls.h
    src/game/GameObjects/GameGeometry/DebugLevel/Walls.c
-
+   src/game/GameObjects/CubeG.h
+   src/game/GameObjects/CubeG.c
     PARENT_SCOPE)
