+++ src/game/Game.c
@@ -1,119 +1,5 @@
 #include "Game.h"
 
-void InitGameObjectManager(GameObjectManager* gameObjectManager)
-{
-	const size_t count = 20u;
-	gameObjectManager->count = count;
-	gameObjectManager->gameObjects = malloc(count * sizeof(GameObject));
-	gameObjectManager->boundingBoxes = malloc(count * sizeof(BoudingBox));
-	gameObjectManager->lastIndex = 0u;
-	gameObjectManager->freeSpace = count;
-
-	if (gameObjectManager->gameObjects == NULL) return;
-	for (size_t i = 0; i < count; i++)
-	{
-		gameObjectManager->gameObjects[i] = malloc(sizeof(GameObject));
-		if (gameObjectManager->gameObjects[i] == NULL) return;
-		InitGameObject(gameObjectManager->gameObjects[i]);
-	}
-}
-
-// TODO FIX THIS METHOD??
-void GameObjectManagerIncrease(GameObjectManager* gameObjectManager)
-{
-	const size_t newCount = gameObjectManager->count + (gameObjectManager->count / 2);
-	GameObject* temp = realloc(gameObjectManager->gameObjects, newCount * sizeof(GameObject));
-	if (temp)
-	{
-		gameObjectManager->gameObjects = temp;
-
-		for (size_t i = gameObjectManager->count; i < newCount; i++)
-		{
-			gameObjectManager->gameObjects[i] = malloc(sizeof(GameObject));
-			if (gameObjectManager->gameObjects[i] == NULL) return;
-			InitGameObject(gameObjectManager->gameObjects[i]);
-		}
-
-		gameObjectManager->count += newCount;
-		gameObjectManager->freeSpace += newCount;
-	}
-}
-
-/*
-When a GameObject is added:
-	add the game object to the end of the game object list at index of count using realloc
-	increase count by 1
- */
-void GameObjectManagerAdd(GameObjectManager* gameObjectManager, GameObject* gameObject)
-{
-	if (gameObjectManager->freeSpace == 0)
-		GameObjectManagerIncrease(gameObjectManager);
-
-	// set its id to the last index
-	gameObject->id = gameObjectManager->lastIndex;
-
-	// set the bounding box
-	gameObject->rigidBody.boundingBox.gameObjectId = gameObjectManager->lastIndex;
-
-
-	// add the address of the bounding box to the array
-	gameObjectManager->boundingBoxes[gameObjectManager->lastIndex] = &gameObject->rigidBody.boundingBox;
-
-	// set the last index to the game object to be added
-	gameObjectManager->gameObjects[gameObjectManager->lastIndex] = gameObject;
-
-	// then call the on start method
-	// NOTE THE ONSTART SHOULD NEVER BE NULL, IF ITS NULL THEN YOU HAVE DONE SOMETHING WRONG, THATS WHY THE IF STATEMENT IS COMMENTED OUT
-	//if (gameObjectManager->gameObjects[gameObjectManager->lastIndex].OnStart != NULL)
-	gameObjectManager->gameObjects[gameObjectManager->lastIndex]->OnStart(gameObjectManager->gameObjects[gameObjectManager->lastIndex]);
-
-
-	// calculate bounding box
-	CalculateBoundingBox(gameObject);
-
-	gameObjectManager->freeSpace--;
-	gameObjectManager->lastIndex++;
-}
-
-/*
-When a GameObject is removed:
-	remove the gameobject from the game object list by setting by freeing it
-	for all gameobjects where index > deleted index
-	copy the gameobject over to gameobject[i - 1]
-	take 1 from count
-	freespace add 1
-	lastIndex take 1
- */
-void GameObjectManagerRemove(GameObjectManager* gameObjectManager, size_t id)
-{
-	FreeGameObject(gameObjectManager->gameObjects[id]);
-	if (gameObjectManager->gameObjects[id] == NULL) return;
-	for (size_t i = id + 1; i < gameObjectManager->count; i++)
-	{
-		// move the game object pointers
-		gameObjectManager->gameObjects[i - 1] = gameObjectManager->gameObjects[i];
-		gameObjectManager->gameObjects[i - 1]->id = i - 1;
-
-		// move the bounding box pointers
-		gameObjectManager->boundingBoxes[i - 1] = gameObjectManager->boundingBoxes[i];
-		// TODO FIX THIS ERORR???
-		gameObjectManager->boundingBoxes[i - 1]->gameObjectId = i - 1;
-
-		// free the last object
-		if (i == gameObjectManager->count - 1)
-			FreeGameObject(gameObjectManager->gameObjects[i]);
-	}
-
-	gameObjectManager->count--;
-	gameObjectManager->freeSpace++;
-	gameObjectManager->lastIndex--;
-}
-
-GameObject* GameObjectManagerFind(GameObjectManager* gameObjectManager, size_t id)
-{
-	return gameObjectManager->gameObjects[id];
-}
-
 void UpdateGameObjects(Time time, GameObjectManager* gameObjectManager)
 {
 	for (size_t i = 0; i < gameObjectManager->lastIndex; i++)
@@ -243,7 +129,7 @@ void FixedUpdateGameObject(Time fixedTime, GameObjectManager* gameObjectManager,
 	//Check for gravity enabled
 	if (gameObject->rigidBody.useGravity && !gameObject->rigidBody.isStatic)
 	{
-		//Apply transformation
+		//Apply gravity transformation
 		GravityTransform(fixedTime, gameObject);
 	}
 
@@ -302,7 +188,7 @@ void GravityTransform(Time fixedTime, GameObject* gameObject)
 	if (rb->sphereBody.isSphere)
 	{
 		// if sphere
-		area = PI * rb->sphereBody.radius;
+		area = M_PI * rb->sphereBody.radius;
 		coefficentDrag = 0.47;
 	}
 	else
@@ -360,24 +246,18 @@ void CalculateBoundingBox(GameObject* gameObject)
 	Transform* transform = &gameObject->transform;
 	Mesh* mesh = &gameObject->mesh;
 
-	Vector3 min = Vec3Multiply(transform->scale, mesh->points[0]);
-	min = Vec3RotateZ(min, transform->rotation.z * PI / 180);
-	min = Vec3RotateY(min, transform->rotation.y * PI / 180);
-	min = Vec3RotateX(min, transform->rotation.x * PI / 180);
-
-
-	min = Vec3Add(transform->position, min);
-
-	Vector3 max = min;
+	Vector3 min;
+	Vector3 max;
 
 	for (size_t i = 0; i < mesh->pointSize; i++)
 	{
 		Vector3 pos = Vec3Multiply(transform->scale, mesh->points[i]);
-		pos = Vec3RotateZ(pos, transform->rotation.z * PI / 180);
-		pos = Vec3RotateY(pos, transform->rotation.y * PI / 180);
-		pos = Vec3RotateX(pos, transform->rotation.x * PI / 180);
-
+		pos = Vec3RotateZ(pos, transform->rotation.z * M_PI / 180);
+		pos = Vec3RotateY(pos, transform->rotation.y * M_PI / 180);
+		pos = Vec3RotateX(pos, transform->rotation.x * M_PI / 180);
 		pos = Vec3Add(transform->position, pos);
+
+		if (i == 0) min = max = pos;
 
 		if (pos.x < min.x) min.x = pos.x;
 		if (pos.y < min.y) min.y = pos.y;
@@ -394,7 +274,16 @@ void CalculateBoundingBox(GameObject* gameObject)
 
 void SphereResolution(Time fixedTime, GameObject* gameObject, GameObject* collidingObject)
 {
-	Vector3 normal = Vec3Normalize(collidingObject->transform.position);
+	Vector3 normal;
+	// assuming that a vector of 0,0,0 is the floor...
+	if (isVec3Empty(collidingObject->transform.position) || collidingObject->rigidBody.isFloor)
+	{
+		normal = (Vector3){ 0.0f, 1.0f, 0.0f };
+	}
+	else
+	{
+		normal = Vec3Normalize(collidingObject->transform.position);
+	}
 
 
 	float dotDirNormal = Vec3DotProduct(gameObject->rigidBody.velocity, normal);
@@ -427,21 +316,14 @@ void SphereResolution(Time fixedTime, GameObject* gameObject, GameObject* collid
 
 
 
-		// transfer half of the decay to the gameobject
-		// add velocity to it
-		// e.g
-		// .25( decay) * mass = veloicty
+		// transfer a % of the decay to the gameobject
+		// .25( decay) * mass / 2 = veloicty
 		if (newDir.x == 0 && newDir.y == 0 && newDir.z == 0) return;
 
 		Vector3 newVel = Vec3ScalarMultiply(Vec3ScalarMultiply(normalNewDir, (decay * collidingObject->rigidBody.mass) / 2), -1.0f);
 		collidingObject->rigidBody.velocity = newVel;
 	}
 
-	// if velocity is less than or equal to if its on the floor, stop
-	//if (Vec3Length(gameObject->rigidBody.velocity) <= 0.7f) return;
-
-
-	//gameObject->rigidBody.velocity = Vec3Multiply(gameObject->rigidBody.velocity, Vec3ScalarMultiply(normalNewDir, decay));
 	if (normalNewDir.x == 1.0f) gameObject->rigidBody.velocity.x *= decay;
 	if (normalNewDir.y == 1.0f) gameObject->rigidBody.velocity.y *= decay;
 	if (normalNewDir.z == 1.0f) gameObject->rigidBody.velocity.z *= decay;
@@ -450,6 +332,8 @@ void SphereResolution(Time fixedTime, GameObject* gameObject, GameObject* collid
 
 }
 
+
+// TODO in code cleanup
 void DetectCollision(Time fixedTime, GameObjectManager* gameObjectManager, GameObject* gameObject)
 {
 	if (gameObject->rigidBody.sphereBody.isSphere)
@@ -505,8 +389,6 @@ void DetectCollision(Time fixedTime, GameObjectManager* gameObjectManager, GameO
 		bool zCollision = objBox->minPos.z <= checkgBox->maxPos.z && objBox->maxPos.z >= checkgBox->minPos.z;
 		if (xCollision && yCollision && zCollision)
 		{
-			//printf("%s is inside of %s\n", gameObject->name, gameObjectManager->gameObjects[checkgBox->gameObjectId]->name);
-
 			SphereResolution(fixedTime, gameObject, gameObjectManager->gameObjects[i]);
 			return;
 		}
@@ -542,6 +424,7 @@ void InitMesh(Mesh* mesh)
 void InitRigidBody(RigidBody* rigidBody)
 {
 	rigidBody->isStatic = false;
+	rigidBody->isFloor = false;
 	rigidBody->useGravity = false;
 	rigidBody->debug = false;
 	rigidBody->mass = 50.0f;
