+++ src/game/GameConstants.h
@@ -1,8 +1,7 @@
 #pragma once
 
+#include "game/GameIncludes.h"
 #include "game/GameTypes.h"
-#include "game/Game.h"
-#include "math/Vector.h"
 
 extern const float MOUSE_SENS;
 extern const float WALK_SPEED;
