+++ src/game/GameIncludes.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <stdlib.h>
+#include <stdio.h>
+#include <stdint.h>
+#include <stdbool.h>
+#include <GL/freeglut.h>
+
+#include "math/Vector.h"
+
+#ifndef M_PI
+#    define M_PI 3.14159265358979323846
+#endif
