+++ src/game/GameObjects/Cube.c
@@ -21,7 +21,7 @@ OnStart OnCubeStart(GameObject* gameObject)
 
 	gameObject->mesh.colors = cubeColorBuffer;
 
-	gameObject->transform.position = (Vector3){ 5.0f, 10.0f, -5.0f };
+	gameObject->transform.position = (Vector3){ 5.0f, 10.0f, 5.0f };
 	gameObject->transform.scale = (Vector3){ 1.0f, 1.0f, 1.0f };
 	gameObject->rigidBody.useGravity = true;
 	gameObject->rigidBody.isStatic = false;
