+++ src/math/mathUtils.h
@@ -1,6 +1,4 @@
 #pragma once
 #include <math.h>
 
-#define PI 3.14159265359
-
 float lerp(float a, float b, float t);
